# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sneak  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-line/1": {
            "put": {
                "tags": [
                    "OrderLines"
                ],
                "summary": "Editar orden-line",
                "description": "Returns order data",
                "operationId": "updateOrderLine",
                "requestBody": {
                    "description": "Ponga sus credenciales",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderLine"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/product": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtiener producto",
                "description": "Obtener todos los productos activos y con stock",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/provider": {
            "get": {
                "tags": [
                    "Providers"
                ],
                "summary": "Obtiener proveedores",
                "description": "Obtener todos los proveedores",
                "operationId": "getProviders",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Provider"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/state": {
            "get": {
                "tags": [
                    "States"
                ],
                "summary": "Obtiener estados",
                "description": "Obtener todos los estados",
                "operationId": "getStates",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/State"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductResource": {
                "title": "ProductResource",
                "description": "Product resource",
                "xml": {
                    "name": "ProductResource"
                }
            },
...
            "OrderLine": {
                "required": [
                    "id, quantity, price, discount, order_id, user_id, product_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "quantity": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "10"
                    },
                    "discount": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "0"
                    },
                    "order_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "user_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "product_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderLine"
                }
            },
...
            "Provider": {
                "required": [
                    "id, name, email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Nike"
                    },
                    "email": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "info@nike.com"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "State"
                }
            },
            "State": {
                "required": [
                    "id, name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "En espera"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "State"
                }
            },
...
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}